feat(ui): support paste-to-attach files on the agent start screen
StartWorkflowView's textarea had no onPaste handler, so users could only
paste files into the live-chat textarea (MessageInput). Drop/upload buttons
worked on both surfaces but pasting was silently dropped on the start
screen.

Extract the shared clipboard-to-File[] conversion (including the
pasted-image renaming heuristic) into a small co-located helper
(clipboardFiles.ts) and add an onPaste handler in StartWorkflowView that
stages the resulting files locally. Callers diverge on what they do with
the resulting File[] — live chat uploads immediately via the parent
onFilesSelected callback; the start screen stages files until the workflow
starts — so the helper covers only the shared extraction step.
